+++ Utils/LoeRound.cs
@@ -2,7 +2,7 @@
 {
   public class LoeRound
   {
-    public int round { get; set; }
+    public ulong round { get; set; }
     public string randomness { get; set; }
     public string signature { get; set; }
   }
